receive: disable bh before using stats seq lock
Otherwise we might get a situation like this:

       CPU0                    CPU1
       ----                    ----
  lock(tstats lock);
                               local_irq_disable();
                               lock(queue lock);
                               lock(tstats lock);
  <Interrupt>
    lock(queue lock);

CPU1 is waiting for CPU0 to release tstats lock. But CPU0, in the
interrupt handler, is waiting for CPU1 to release queue lock. The
solution is to disable interrupts on CPU0, so that this can't happen.

Note that this only affects 32-bit, since u64_stats_update_begin nops
out on native 64-bit platforms.

Reported-by: René van Dorst <opensource@vdorst.com>
Signed-off-by: Jason A. Donenfeld <Jason@zx2c4.com>