compat: don't clear header bits on RHEL
The 3.10 kernel from Red Hat puts head_frag in between headers_start and
headers_end. This is triggered on the latest i40e driver. This results
in the packet being freed incorrectly, crashing the system. So, this
patch just ensures we don't zero any of the header bits.

The whole issue of zeroing header bits probably should be revisited
sometime somewhat soon.

Signed-off-by: Jason A. Donenfeld <Jason@zx2c4.com>